Seperated db connection to a seperate module Added basic get for API routes
